ODMS/requirements.txt
appdirs==1.4.4
bidict==0.21.2
black==21.7b0
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
passlib==1.7.4
pathspec==0.9.0
protobuf==3.17.3
python-engineio==4.2.0
python-socketio==5.3.0
regex==2021.7.6
six==1.16.0
stdiomask==0.0.6
tomli==1.0.4
Werkzeug==2.0.1
